Move Buzz\Adapter to Buzz\Client
