+++ lib/Buzz/Client/AbstractAdapter.php
@@ -1,10 +1,9 @@
 <?php
 
-namespace Buzz\Adapter;
+namespace Buzz\Client;
 
 use Buzz\Message\RequestInterface;
 use Buzz\Message\MessageInterface;
-use Buzz\Client\ClientInterface;
 
 abstract class AbstractAdapter implements ClientInterface
 {
